+++ src/errors.ts
@@ -1,0 +1,7 @@
+export class InvalidInputError extends Error {
+    constructor(message = 'Invalid input') {
+        super(message);
+        this.name = 'InvalidInputError';
+        Object.setPrototypeOf(this, InvalidInputError.prototype);
+    }
+}